Contenidos/05_ia_ds_modulo_5/03_clases_practicas/04_Clase_04/app/requirements.txt
streamlit>=1.10
google-genai
python-dotenv